requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
groq>=0.4.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
langgraph>=0.0.30
pydantic>=2.0.0
python-docx>=1.1.0
